Add Template parsing
This adds basic template parsing without using reflection to try
and provide somewhat more sane debug output in the case of typos
or misuse. We also made the plan yaml an array instead of a map
because the yaml parser under the hood uses an unordered map.
We need to run these in order so that doesn't work at all and
doing anything else to enforce order felt very ugly.