Refactor Provider Code for Templates
Not going to lie... tried to make a nice interface for the providers
but go isn't being super friendly to me right now. This is mostly because
I need to work with go templates and in order to merge several providers
i need to turn everything into an interface. Since that is the only time
i'm really using the providers it makes zero sense to spend time parsing
them into nice structs just to rip them apart later and place them in a
map.